More

+++ apps/webserver/src/index.js
@@ -52,6 +52,9 @@ app.post('/upload', (req, res, next) => {
     let number;
     let mode;
     let rep;
+    let r;
+    let s;
+    let a;
 
     req.busboy.on('field', (fieldName, value) => {
         if(fieldName === 'precision') {
@@ -68,6 +71,21 @@ app.post('/upload', (req, res, next) => {
             rep = value;
             console.log('Repeat set at', rep);
         }
+
+        if(fieldName === 'inputSize') {
+            r = value;
+            console.log(`Input resized to ${r}px wide`);
+        }
+
+        if(fieldName === 'outputSize') {
+            s = value;
+            console.log(`Output will be resized to ${s}px wide`);
+        }
+
+        if(fieldName === 'alpha') {
+            a = value;
+            console.log('Alpha set to', a);
+        }
     });
 
     req.busboy.on('file', (field, file, fileName) => {
@@ -75,8 +93,8 @@ app.post('/upload', (req, res, next) => {
         file.pipe(fstream);
         fstream.on('close', () => {
             console.info('Uploaded', fileName);
-            child.exec(`primitive -i ./dist/static/input/${fileName} -o ./dist/static/output/${fileName} -n ${number || '100'} -m ${mode || '0'} -rep ${rep || '0'}`, (err, stdout, stderr) => {
-                if(!err) {
+            child.exec(`primitive -i ./dist/static/input/${fileName} -o ./dist/static/output/${fileName} -n ${number || '100'} -m ${mode || '0'} -rep ${rep || '0'} -r ${r || '256'} -s ${s || '1024'} -a ${a || '128'}`, (err, stdout, stderr) => {
+                if(!err && !stderr) {
                     console.log('Rendered with precision of', number);
                     fs.emptyDir(`${__dirname}/static/input`, (err) => {
                         if(!err) {
@@ -87,7 +105,12 @@ app.post('/upload', (req, res, next) => {
                         }
                     })
                 } else {
-                    console.error(err);
+                    if(err) {
+                        console.error(err);
+                    }
+                    if(stderr) {
+                        console.error(stderr);
+                    }
                 }
             });
         });
